% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchangr.R
\docType{package}
\name{exchangr}
\alias{exchangr}
\alias{exchangr-package}
\title{exchangr: A package for resampling data from complex designs.}
\description{
The package exchangr currently provides functions to randomly
relabel observations in a way that respects design constraints.
}
