% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{has_pseudoreplications}
\alias{has_pseudoreplications}
\title{Does the dataset have pseudoreplications for a given factor?}
\usage{
has_pseudoreplications(data, x, unit)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Unquoted name of variable to check (must be character or factor).}

\item{unit}{Unquoted name of variable identifying individual sampling units.}
}
\description{
Does the dataset have pseudoreplications for a given factor?
}
\details{
Tests whether there are multiple observations per level of
  factor \code{x} per unit.
}
