% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_within}
\alias{is_within}
\title{Check whether levels of a factor are within- or between- unit.}
\usage{
is_within(data, x, unit)
}
\arguments{
\item{data}{Unquoted name of a data frame.}

\item{x}{Unquoted name of variable to check.}

\item{unit}{Unquoted name of sampling unit factor.}
}
\description{
Check whether levels of a factor are within- or between- unit.
}
\details{
First checks whether the factor \code{x} is between unit
  by testing that each sampling unit has only a single level. If
  not, then the factor is assumed to be within. The user is warned
  if some units are missing levels. This behavior can be turned off
  by wrapping the call in \code{suppressWarnings()}.
}
\examples{
d <- data.frame(
  subject = rep(1:2, each = 2),
  A = rep(paste0("A", 1:2), 2),
  B = rep(paste0("B", 1:2), each = 2))
}
