% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{shuffling-functions}
\alias{shuffling-functions}
\alias{shuffle}
\alias{shuffle_sync}
\alias{shuffle_each}
\alias{shuffle_each_sync}
\title{Basic functions for shuffling labels}
\usage{
shuffle(.data, x)

shuffle_sync(.data, x, ...)

shuffle_each(.data, x, unit)

shuffle_each_sync(.data, x, unit, ...)
}
\arguments{
\item{.data}{Data frame.}

\item{x}{Unquoted name of factor whose levels are to be shuffled.}

\item{...}{Unquoted names of factors over whose combinations the
shuffling is to be synchronized.}

\item{unit}{Unquoted name of variable within the levels of which
the shuffling of 'x' is to be restricted.}
}
\description{
These functions reflect alternative ways of shuffling labels so as
to respect the constraints of the study design.
}
\examples{
# Use shuffle for data with a single between subjects factor
d <- data.frame(group = rep(c("treatment", "control"), c(6, 6)),
                subject = 1:12)
shuffle(d, group)

shuffle_sync(three_way, A, B, C) # shuffle A over B-C
shuffle_sync(three_way, B, A, C) # shuffle B over A-C

# For multilevel data, nest any within-subject factors
# before shuffling. There should be one row per subject.
three_nest <- tidyr::nest(three_mix, A)
d <- shuffle_sync(three_nest, B, C)
# then restore
tidyr::unnest(d)

\dontrun{
# throws error because within-subject factor B introduced
# pseudoreplications
shuffle_each(two_within, A, subject)
}
# nest first
nested <- tidyr::nest(two_within, B)
result <- shuffle_each(nested, A, subject)
# then restore
tidyr::unnest(result)

shuffle_each_sync(three_mix, A, subject, B) # over B, ignore C
shuffle_each_sync(three_mix, A, subject, B, C) # over B:C
}
