% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{three_mix}
\alias{three_mix}
\title{Example data from a 3x2x2 mixed design}
\format{A data frame with 24 rows and four variables:
\describe{
  \item{subject}{Subject identifier.}
  \item{A}{Three level within subjects factor (A1, A2, A3).}
  \item{B}{Two level between subjects factor (B1, B2).}
  \item{C}{Two level between subjects factor (C1, C2).}
}}
\usage{
three_mix
}
\description{
A data frame containing labels corresponding to 24 observations
from a balanced 3x2x2 full factorial design, with one within and
two between factors.
}
\keyword{datasets}
