% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{three_way}
\alias{three_way}
\title{Example data from a 3x2x2 between-subjects design}
\format{A data frame with 24 rows and three variables:
\describe{
  \item{subject}{Subject identifier.}
  \item{A}{Three level factor (A1, A2, A3).}
  \item{B}{Two level factor (B1, B2).}
  \item{C}{Two level factor (C1, C2).}
}}
\usage{
three_way
}
\description{
A data frame containing labels corresponding to 24 observations
from a balanced 3x2x2 full factorial design, with all factors
between subjects.
}
\keyword{datasets}
