% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_within}
\alias{two_within}
\title{Example data from a 2x2 within-subjects design}
\format{A data frame with 24 rows and three variables:
\describe{
  \item{subject}{Subject identifier.}
  \item{A}{Two level factor (A1, A2).}
  \item{B}{Two level factor (B1, B2).}
}}
\usage{
two_within
}
\description{
A data frame containing labels corresponding to 24 observations
from a balanced 2x2 full factorial design, with both factors within
subjects.
}
\keyword{datasets}
